refactor(streaming): move streaming_utils to streaming package, update imports, update plan
